old_requirements.txt
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
idna==3.8
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
MouseInfo==0.1.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32==306
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
